V.0.7.1
I worked hard to provide this version.
I'll enhance the .html file (cause its ugly) or you can simple provide me some cool designs.
Hope it will be easier for you to run it.
It will provide the list of the files needed if you only want to run it on a server.